pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-intel"
version = "1.0.0"
description = "AI Video Intelligence Platform — Netflix X-Ray style face & object detection"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Video Intelligence Team" }]

dependencies = [
    "fastapi==0.95.2",
    "uvicorn==0.22.0",
    "celery==5.3.0",
    "redis==5.0.0",
    "psycopg2-binary==2.9.6",
    "pgvector==0.1.0",
    "opencv-python-headless==4.8.0.76",
    "insightface==0.7.3",
    "numpy==1.26.4",
    "onnxruntime==1.19.2",
    "ultralytics==8.4.19",
    "deep-sort-realtime==1.3.2",
    "torch==2.4.0",
    "torchvision==0.19.0",
    "python-dotenv",
    "Pillow",
    "datasets>=2.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",          # HTTPX test client for FastAPI
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
